requirements.txt
addict==2.4.0
albumentations==1.3.1
asttokens==2.4.1
backcall==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
efficientnet-pytorch==0.7.1
executing==2.0.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mlxtend==0.23.0
mmcv==1.6.1
munch==4.0.0
networkx==3.1
numpy==1.24.4
opencv-python==4.6.0.66
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.1.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.41
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.0
six==1.16.0
stack-data==0.6.3
tensorboardX==2.5.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
torch==1.10.0
torch-dct==0.1.6
torch-utils==0.1.2
torchvision==0.11.0
tqdm==4.66.1
traitlets==5.14.0
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
yapf==0.40.1
zipp==3.17.0
